modified:   workspaces/DummyRosterRC0.1/DummyRoster.Common.EntityModel/Models/Employee.cs

+++ workspaces/DummyRosterRC0.1/DummyRoster.Common.EntityModel/Models/Employee.cs
@@ -9,7 +9,7 @@ namespace DummyRoster.Common.EntityModel.Models;
 public partial class Employee
 {
   [Key]
-  public long Id { get; set; }
+  public int Id { get; set; }
 
   [Column(TypeName = "varchar(32)")]
   public string Name { get; set; } = null!;
@@ -28,10 +28,10 @@ public partial class Employee
   public string? Belonging { get; set; }
 
   [Column(TypeName = "INT")]
-  public long? Loc { get; set; }
+  public int? Loc { get; set; }
 
   [Column(TypeName = "INT")]
-  public long? Ref { get; set; }
+  public int? Ref { get; set; }
 
   [InverseProperty("ContactNavigation")]
   public virtual ICollection<Carrier> Carriers { get; } = new List<Carrier>();
